Referring to Issue #119
The colour schema on the pie chart isn't that accurate in nature. 
Hence a new colour schema has been provided.

The issue with the old colour schema was that beyond a certain number of values the thing got looped and the same colours were used over and over again. Hence making a colour schema which is accurate was of utmost importance and making this thing usable to everyone in a very simpler way. 

Note to devs - Do lemme know what do think about it